added version, modified package
The script is now packaged as a real library (only holds the version) and
the script (still where all the good stuff is at). But this now means that
I can start refactoring to pull stuff out. So good for me (pats self on
back). Also,  now does something and the version is set in one place